% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionfeed.R
\name{buttonkrmesproductionfeedodsbillUI_right}
\alias{buttonkrmesproductionfeedodsbillUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionfeedodsbillUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
