% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionorder.R
\name{buttonkrmesproductionorderodsbillUI_right}
\alias{buttonkrmesproductionorderodsbillUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionorderodsbillUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
