% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionorder.R
\name{buttonkrmesproductionorderodsUI_right}
\alias{buttonkrmesproductionorderodsUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionorderodsUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
