% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionrequisition.R
\name{buttonkrmesproductionrequisitionodsbillUI_left}
\alias{buttonkrmesproductionrequisitionodsbillUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionrequisitionodsbillUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonsalaryUI_left()
}
