% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionrequisition.R
\name{buttonkrmesproductionrequisitionodsUI_left}
\alias{buttonkrmesproductionrequisitionodsUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionrequisitionodsUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonsalaryUI_left()
}
