% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionrequisition.R
\name{buttonkrmesproductionrequisitionodsUI_right}
\alias{buttonkrmesproductionrequisitionodsUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionrequisitionodsUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonsalaryUI_right()
}
