% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionwarehousin.R
\name{buttonkrmesproductionwarehousinodsbillUI_right}
\alias{buttonkrmesproductionwarehousinodsbillUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionwarehousinodsbillUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
