% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionwarehousin.R
\name{buttonkrmesproductionwarehousinodsUI_left}
\alias{buttonkrmesproductionwarehousinodsUI_left}
\title{操作区域}
\usage{
buttonkrmesproductionwarehousinodsUI_left()
}
\value{
返回值
}
\description{
操作区域
}
\examples{
buttonUI_left()
}
