% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionwarehousin.R
\name{buttonkrmesproductionwarehousinodsUI_right}
\alias{buttonkrmesproductionwarehousinodsUI_right}
\title{预览区域}
\usage{
buttonkrmesproductionwarehousinodsUI_right()
}
\value{
返回值
}
\description{
预览区域
}
\examples{
buttonUI_right()
}
