% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R, R/jhhrv.R
\name{buttonUI_bottom}
\alias{buttonUI_bottom}
\title{生成区域}
\usage{
buttonUI_bottom()

buttonUI_bottom()
}
\value{
返回值

返回值
}
\description{
生成区域

生成区域
}
\examples{
buttonUI_bottom()
buttonUI_bottom()
}
