% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhhrv.R
\name{jhhrvUI}
\alias{jhhrvUI}
\title{按纽生成生成器界面}
\usage{
jhhrvUI(
  tabTitle = "嘉好人资凭证区域",
  colTitles = c("上传数据", "生成凭证", "显示区域"),
  widthRates = c(6, 6, 12),
  func_left = buttonUI_left,
  func_right = buttonUI_right,
  func_bottom = buttonUI_bottom
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
buttonUI()
}
