% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesemp.R
\name{krmesempodsUI}
\alias{krmesempodsUI}
\title{按纽生成生成器界面}
\usage{
krmesempodsUI(
  tabTitle = "员工",
  colTitles = c("操作区", "显示区域"),
  widthRates = c(4, 8),
  func_left = buttonkrmesempodsUI_left,
  func_right = buttonkrmesempodsUI_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
krmesempodsUI()
}
