% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krmesproductionrequisition.R
\name{krmesproductionrequisitionodsUI}
\alias{krmesproductionrequisitionodsUI}
\title{按纽生成生成器界面}
\usage{
krmesproductionrequisitionodsUI(
  tabTitle = "生产领料单",
  colTitles = c("操作区", "显示区域"),
  widthRates = c(4, 8),
  func_left = buttonkrmesproductionrequisitionodsUI_left,
  func_right = buttonkrmesproductionrequisitionodsUI_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{colTitles}{主页标题}

\item{widthRates}{左右比例}

\item{func_left}{左函数}

\item{func_right}{右函数}

\item{func_bottom}{下面一栏}
}
\value{
返回值
}
\description{
按纽生成生成器界面
}
\examples{
jhhrvodssalaryUI()
}
